requirement.txt
feedparser==6.0.11
PyYAML==6.0.2
Jinja2==3.1.4